+++ vpin-studio-ui/src/main/java/de/mephisto/vpin/ui/tables/TableOverviewController.java
@@ -1121,7 +1121,7 @@ public class TableOverviewController extends BaseTableController<GameRepresentat
       if (backup != null) {
         Button compBtn = new Button();
         compBtn.getStyleClass().add("table-media-button");
-        compBtn.setTooltip(new Tooltip("A backup for this table has been found"));
+        compBtn.setTooltip(new Tooltip("Show the backup of this table in the backups view."));
         FontIcon cmpIcon = WidgetFactory.createIcon("mdi2a-archive-outline");
         compBtn.setGraphic(cmpIcon);
         row.getChildren().add(compBtn);