+++ RELEASE_NOTES.md
@@ -40,3 +40,4 @@
 ## Bugfixes
 
 - Fixed wrong update check that happened when one client was connected/disconnected to different servers.
+- Fixed pause menu that did not load the highscore card.